requirements.txt
fastapi
uvicorn
langchain
langchain-community
langgraph
openai
langchain_openai
sqlalchemy
pymysql
python-dotenv
